opt: move queue module out of netpoll module
