+++ internal/netpoll/epoll.go
@@ -32,7 +32,7 @@ import (
 	"golang.org/x/sys/unix"
 
 	"github.com/panjf2000/gnet/errors"
-	"github.com/panjf2000/gnet/internal/netpoll/queue"
+	"github.com/panjf2000/gnet/internal/queue"
 	"github.com/panjf2000/gnet/logging"
 )
