+++ internal/queue/queue_test.go
@@ -5,7 +5,7 @@ import (
 	"sync/atomic"
 	"testing"
 
-	"github.com/panjf2000/gnet/internal/netpoll/queue"
+	"github.com/panjf2000/gnet/internal/queue"
 )
 
 func TestLockFreeQueue(t *testing.T) {
